test(entities-upstreams-targets): fix flaky tests and test results upload workflow (#714)
The `cy.wait()` will wait intercepted network events for 5 seconds
by default. But in the CI environment, the
`['@getUpstream', '@fetchServices', '@fetchCertificates']`
combination will took more than 4 seconds so the test will be flaky
if it occasionally execeeds the 5 seconds limitation.

The slow part seems caused by the Cypress's vm hooks. It does not
affect the component in development/production mode. So it's safe
to loose the timeout to 10 seconds

This commit also fix the test results upload failure. The artifact
name could not contains some special characters like `/`, which
will occur in `matrix.container`. A new step is added to normalize
the artifact name to make sure it can be uploaded successfully.