fix(tsconfig): apply `@vue/tsconfig`
Use an exported `const enum` in vite is forbidden due to the
[compiler limitation](https://vitejs.dev/guide/features.html#isolatedmodules).
To fix this, some `const enum` exports need to be reshape to `enum`.

Besides the `const enum`, we also need to use the Type-Only Imports
and Export syntax to avoid potential problems like type-only imports
being incorrectly bundled. See https://vitejs.dev/guide/features.html#transpile-only

These restrictions are all included in `@vue/tsconfig`, so we applied
it to the root `tsconfig.json` to prevent further issues.