fix: do not inherit height and width for internal chart parent
the height that is set on the shell, should be the maximum allowed
hight the chart should have internally. Setting that to inherit
would cause the chart internally to overflow its container

Lets say the chart shell had a max hight of 500px

The chart parent which tells chartjs how large the chart should be
would also try to be 500px.

500px is what the whole chart shell, including legend, header, etc should have so
the result is that the chart would overflow its parent container.

if we set the