feat(plugins): allow creating dup plugins [khcp-10683] (#1153)
Remove the logic that prevents users from adding duplicate plugins of the same type to the same entity. We need to support creating multiple plugins of the same type for `consumers`/`consumer_groups` to allow scoping to multiple entities simultaneously (ie. I can create a plugin scoped to a specific `consumer_group`/`route` and have a different configuration for the same plugin when it's scoped to a specific `consumer_group`/`gateway-service` correctly).
Addresses [KHCP-10683](https://konghq.atlassian.net/browse/KHCP-10683).