fix: some styling issues/size calculations
dynamically set the hight of the horizontal legend based on if there needs to be
more than one row or not.

Only need to allocate space for one row if there's only one row.

If there's more than one row, allocate space for 2 rows, the rest overflows.

Set the chart-container hight to 100% - 80px to account for the legend + chart title
when chart and legend is stacked