refactor(cypress): better typings for `cy.mount` and `cy.get('@vueWrapper')`
This PR improves the typings for `cy.mount` and `cy.get('@vueWrapper')`,
making it more accurate and able to detect the wrong props usage.

As `cy.get('@vueWrapper')` is widely used across the project, the
chainable typing is specificed to reflect the real value inside the wrapper.

Some obvious wrong props are fixed accordingly in this PR

JIRA: KM-366